Update debug.md - logger level debug no longer needed
logger level debug is no longer required according to this post
https://community.home-assistant.io/t/new-debug-component-how-to-get-reset-reason-information-seperate/393754/10

and pr:
https://github.com/esphome/esphome/pull/8980#issuecomment-2924511024

I have removed this and added some text explaining that it will also show some details in the web server including a link to it in the text and in see also (I think, preview does not show right, but it seems to be how the others are done). 

hope this helps.